Add files via upload
Excel Project of Veena Stores Sales Analysis
In this Excel project the stores data is used in order to data cleaning, processing, analysis and dashboard creation in Excel. With the help of Pivots analysis is being made on what all parameters have to be considered in order to increase the sales.